Always index Jack ports and refer to them by name
The Jack API annoyingly doesn't provide a way to get an ID from a port name.
Since notification callbacks apparently don't have to worry about realtime
concerns anymore, simply use the full name string as an ID everywhere.

This means that every Jack port has a non-null ID, and all ports are always
indexed, so the kludges for this in the canvas can be removed.