Abstract out sending of events
This removes the details of how events are handled from drivers, so the owner
can set them up to do anything.  For example, a driver could be run in the GUI
thread and have its events simply dispatched immediately, but here everything
is enqueued to the same queue which is drained later for simplicity.