Remove fuzzy markers from translation files
As I understand it, these are a signal from xgettext that it has changed an
entry, so human revision may be required.  The files checked in to the source
tree have been human reviewed, and so should not have any "fuzz" in them.

This wouldn't really be an issue if meson/xgettext didn't annoyingly smash
patchage.pot when .po files are updated, but such is life.